requirements.txt
python-multipart

fastapi
uvicorn
easyocr
pillow
numpy